more moving files around, fix tests
