+++ .github/workflows/test.yaml
@@ -17,5 +17,5 @@ jobs:
           architecture: "x64" # optional x64 or x86. Defaults to x64 if not specified
       - run: python -m pip install -e cioos_data_transform
       - run: python -m pip install -e projects
-      - working-directory: projects/odf_transform
-        run: python test_odf.py
+      - working-directory: projects/odf_transform/odf_transform
+        run: python odf_to_netcdf.py
