+++ projects/odf_transform/odf_transform/vocabulary/README.md
@@ -1,3 +1,3 @@
 # Vocabulary
 
-This list is built from the files in the vocabulary folder, see all_vocabulary_list.csv
+This list is built from the files in the vocabulary folder, see ../vocabulary_source_files/vocabulary_list.csv
